requirements.txt
scikit-learn
pillow
numpy
